Refactor out send_court_warning_letter classification
